Generalize MethodAndClass to support other languages
